% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nic_colours.R
\docType{data}
\name{nic_palettes}
\alias{nic_palettes}
\title{Nature Inspired Colour Palatte Package}
\format{
An object of class \code{list} of length 15.
}
\usage{
nic_palettes
}
\description{
This is a collection of color palettes for R inspired by leaves. This work is inspired by the
pnwcolours package in R and codes are modifies using pnwcolours.
Create colour palettes
Names: applecroton, colleasa, colleasb
Type  \code{names(nic_palettes)} to view the names of colour palattes
}
\keyword{datasets}
